Moved calculate reward method to RLLearner and adjusted RLLearner class by adding variables, innit
